backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
